robot_vision_trainer/requirements.txt
# Robot Vision Trainer - Combined Service Requirements
requests==2.32.3
flask==2.3.3
pillow==10.0.0
python-dotenv==1.0.0
tqdm==4.67.1
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.36.0
huggingface_hub>=0.19.4
datasets>=2.15.0
lerobot>=0.1.0
accelerate>=0.25.0
wandb>=0.16.0
pyyaml>=6.0.1
scikit-learn>=1.3.2
openai>=1.13.0
python-multipart>=0.0.7 